Finishes blockchain logic
