+++ app/Model/Transaction.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel
 
 
 class Transaction(BaseModel):
-    id: str = str(uuid.uuid4())[0:8]
+    id: str = str(uuid.uuid4())
     timestamp: int = time.time_ns()
     sender_wallet: Union[str, None]
     recipient: Union[str, None]
