+++ app/Model/Wallet.py
@@ -10,9 +10,8 @@ class Wallet:
     Keeps track of the mine's balance.
     Allows a miner to authorizse transactions
     """
-    # TODO: Invesigar
     def __init__(self):
-        self.address = str(uuid.uuid4())[0:8]
+        self.address = str(uuid.uuid4())
         self.private_key = ec.generate_private_key(ec.SECP256K1(), default_backend())
         self.public_key = self.private_key.public_key()
         self.my_credits = []
